backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
python-decouple==3.8
asgiref==3.9.1
sqlparse==0.5.3
requests==2.32.5
drf-spectacular==0.28.0 